backend/services/ResumeService/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
PyMuPDF==1.23.6
python-docx==1.1.0
python-multipart==0.0.6
python-dotenv==1.0.0